+++ app/src/main/java/com/yapp/chaeum/ui/AppState.kt
@@ -1,22 +1,18 @@
 package com.yapp.chaeum.ui
 
-import android.annotation.SuppressLint
-import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.navigation.NavDestination
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
+import com.example.auth.navigation.AuthGraphDest
 import com.example.etc.navigation.navigateToEtc
-import com.example.matching.navigation.MatchingGraph
-import com.example.matching.navigation.navigateToMatching
+import com.example.matching.navigation.navigateToMatchingGraph
 import com.example.mypage.navigation.navigateToMyPage
-import com.yapp.chaeum.navigation.HomeGraph
 import com.yapp.chaeum.navigation.TopLevelDestination
+import com.yapp.chaeum.navigation.home.HomeGraph
 
 @Composable
 fun rememberAppState(
@@ -34,29 +30,29 @@ class AppState(
         @Composable get() = navController
             .currentBackStackEntryAsState().value?.destination
 
+    fun navigateToHome() {
+        navController.navigate(HomeGraph) {
+            popUpTo(AuthGraphDest.AuthRoute) {
+                inclusive = true
+            }
+            launchSingleTop = true
+            restoreState = true
+        }
+    }
+
     fun navigateToTopLevelDestination(
         topLevelDestination: TopLevelDestination,
     ) {
-        Log.d(
-            "whatisthis",
-            "topLevelNavOptions ${navController.graph.findStartDestination().id}, $topLevelDestination"
-        )
         val topLevelNavOptions = navOptions {
-            // Pop up to the start destination of the graph to
-            // avoid building up a large stack of destinations
-            // on the back stack as users select items
             popUpTo(HomeGraph) {
                 saveState = true
             }
-            // Avoid multiple copies of the same destination when
-            // reselecting the same item
             launchSingleTop = true
-            // Restore state when reselecting a previously selected item
             restoreState = true
         }
 
         when (topLevelDestination) {
-            TopLevelDestination.MATCHING -> navController.navigateToMatching(topLevelNavOptions)
+            TopLevelDestination.MATCHING -> navController.navigateToMatchingGraph(topLevelNavOptions)
             TopLevelDestination.MYPAGE -> navController.navigateToMyPage(topLevelNavOptions)
             TopLevelDestination.ETC -> navController.navigateToEtc(topLevelNavOptions)
         }
