+++ feature/matching/src/main/java/com/example/matching/detail/MatchingDetailRoute.kt
@@ -1,5 +1,6 @@
-package com.example.etc
+package com.example.matching.detail
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Text
@@ -9,14 +10,16 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.sp
 
 @Composable
-fun EtcRoute(
+fun MatchingDetailRoute(
+    onBack: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    Box(modifier = modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text(text = "EtcRoute", fontSize = 30.sp)
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+            .clickable { onBack() },
+        contentAlignment = Alignment.Center
+    ) {
+        Text(text = "MatchingDetailRoute", fontSize = 30.sp)
     }
-}
-
-@Composable
-internal fun EtcScreen() {
 }
